Free-telegram-bot-monitoring-OKX
This is my first program.
This program is to monitor the price change of stocks from OKS-web site using telegram bot.
In order to get started you need to create a telegram bot and fill in the config file (token and chat_id). Next step is to run controller.py and open telegram chat.